refactor: move field helpers to helpers.js
